requirements.txt
python-dotenv
pymongo
pandas
numpy
scikit-learn
certifi
pymongo[srv]
pymongo
scikit-learn
scipy
dill
pyaml

#-e .